requirements.txt
africastalking==1.2.9
alembic==1.12.0
aniso8601==9.0.1
apturl==0.5.2
Authlib==1.3.2
bcrypt==3.2.0
blinker==1.6.2
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
coverage==7.6.7
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.2.2
Faker==19.11.0
fasteners==0.14.1
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
future==0.18.2
greenlet==3.0.0
gunicorn==20.1.0
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==24.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipenv==2023.9.8
platformdirs==3.11.0
pluggy==1.5.0
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==8.3.3
pytest-cov==6.0.0
pytest-flask==1.3.0
pytest-mock==3.10.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.32.3
schema==0.7.7
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.21
SQLAlchemy-serializer==1.4.1
systemd-python==234
tomli==2.1.0
typing_extensions==4.8.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.24.5
virtualenv-clone==0.3.0
wadllib==1.3.6
Werkzeug==3.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
